refactor : 이벤트 인터페이스 및 payload 패키지 구조 개선
