+++ apps/asset-service/src/main/java/io/antcamp/assetservice/domain/event/payload/CompetitionEndedEvent.java
@@ -1,4 +1,4 @@
-package io.antcamp.assetservice.infrastructure.messaging.kafka.payload;
+package io.antcamp.assetservice.domain.event.payload;
 
 import java.time.LocalDateTime;
 import java.util.List;
