+++ apps/asset-service/src/main/java/io/antcamp/assetservice/infrastructure/messaging/kafka/consumer/CompetitionEndedEventConsumer.java
@@ -6,9 +6,9 @@ import io.antcamp.assetservice.domain.model.Account;
 import io.antcamp.assetservice.domain.model.Holding;
 import io.antcamp.assetservice.domain.repository.AccountRepository;
 import io.antcamp.assetservice.domain.repository.HoldingRepository;
-import io.antcamp.assetservice.domain.repository.TotalAssetEventProducer;
+import io.antcamp.assetservice.application.event.TotalAssetEventProducer;
 import io.antcamp.assetservice.infrastructure.client.StockPriceClient;
-import io.antcamp.assetservice.infrastructure.messaging.kafka.payload.CompetitionEndedEvent;
+import io.antcamp.assetservice.domain.event.payload.CompetitionEndedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
